Move channel into the subscription
Similar to ocm-ramen-sample PR[1] move the channel into the subscription
kustomization.

This is better for users, removing setup and teardown from the
deployment steps.  Now it is not possible to forget to deploy or remove
the channel after testing.

This is better for developers, since now we can test multiple VMs
concurrently, without dealing with setup and cleanup of shared
resources, since VMs do not share the channel.

[1] https://github.com/RamenDR/ocm-ramen-samples/pull/41